# Docs - contract/openapi.yaml
openapi: 3.0.1
info:
  title: Homely API
  description: HTTP JSON dùng {statusCode,message,data}; DTO trong data dùng snake_case, VND là chuỗi số nguyên. Thanh toán và hoàn tiền là sandbox.
  version: 1.6.0
servers:
- description: Local backend; paths include /api/v1
  url: http://localhost:8080
paths:
...
  /api/v1/bookings:
    post:
      tags:
      - booking-controller
      summary: Create booking request
      operationId: createBooking
      parameters:
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/BookingDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/bookings/{id}/reject:
    post:
      tags:
      - booking-controller
      summary: Reject booking
      operationId: rejectBooking
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingActionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/BookingDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/bookings/{id}/handover-tenant:
    post:
      tags:
      - booking-controller
      summary: Tenant confirms handover
      operationId: tenantHandover
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/BookingDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/bookings/{id}/handover-host:
    post:
      tags:
      - booking-controller
      summary: Host confirms handover
      operationId: hostHandover
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/BookingDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/bookings/{id}/confirm-deposit:
    post:
      tags:
      - booking-controller
      summary: Confirm deposit payment
      operationId: confirmDeposit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/BookingDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/bookings/{id}/cases:
    get:
      tags:
      - booking-controller
      operationId: cases
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookingCaseDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
    post:
      tags:
      - booking-controller
      operationId: openCase
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseCreateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/BookingCaseDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/bookings/{id}/cancel:
    post:
      tags:
      - booking-controller
      summary: Cancel booking
      operationId: cancelBooking
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingActionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/BookingDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/bookings/{id}/approve:
    post:
      tags:
      - booking-controller
      summary: Approve booking
      operationId: approveBooking
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/BookingDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
  /api/v1/payments/booking/{bookingId}:
    get:
      tags:
      - payment-controller
      summary: Get payment info for booking
      operationId: getPaymentInfo
      parameters:
      - name: bookingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/PaymentDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
  /api/v1/bookings/{id}:
    get:
      tags:
      - booking-controller
      operationId: detail_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/BookingDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/bookings/my:
    get:
      tags:
      - booking-controller
      summary: Get my bookings
      operationId: getMyBookings
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/PageResponseBookingDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/bookings/host:
    get:
      tags:
      - booking-controller
      summary: Get host bookings
      operationId: getHostBookings
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/PageResponseBookingDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
components:
  schemas:
...
    PaymentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        provider:
          type: string
        status:
          type: string
        booking_id:
          type: integer
          format: int64
        payer_id:
          type: integer
          format: int64
        provider_payment_id:
          type: string
        amount_vnd:
          pattern: ^[0-9]+$
          type: string
        expires_at:
          type: string
          format: date-time
        succeeded_at:
          type: string
          format: date-time
        is_test:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    BookingCreateRequest:
      required:
      - room_id
      type: object
      properties:
        note:
          maxLength: 1000
          minLength: 0
          type: string
        room_id:
          type: integer
          format: int64
        desired_move_in:
          type: string
          format: date
        occupant_count:
          maximum: 10
          minimum: 1
          type: integer
          format: int32
    BookingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
        note:
          type: string
        version:
          type: integer
          format: int32
        tenant_id:
          type: integer
          format: int64
        host_id:
          type: integer
          format: int64
        room_id:
          type: integer
          format: int64
        listing_id:
          type: integer
          format: int64
        desired_move_in:
          type: string
          format: date
        occupant_count:
          type: integer
          format: int32
        rent_vnd:
          pattern: ^[0-9]+$
          type: string
        deposit_vnd:
          pattern: ^[0-9]+$
          type: string
        request_expires_at:
          type: string
          format: date-time
        hold_expires_at:
          type: string
          format: date-time
        handover_due_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        cancelled_at:
          type: string
          format: date-time
        last_reason:
          type: string
        created_at:
          type: string
          format: date-time
    BookingActionRequest:
      type: object
      properties:
        reason:
          maxLength: 1000
          minLength: 0
          type: string
    CaseCreateRequest:
      required:
      - description
      - type
      type: object
      properties:
        type:
          type: string
          enum:
          - DEPOSIT_DISPUTE
          - HANDOVER_OVERDUE
          - NO_SHOW
          - OTHER
        description:
          maxLength: 2000
          minLength: 0
          type: string
    BookingCaseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        status:
          type: string
        description:
          type: string
        decision:
          type: string
        booking_id:
          type: integer
          format: int64
        resolution_note:
          type: string
        tenant_refund_vnd:
          pattern: ^[0-9]+$
          type: string
        host_retained_vnd:
          pattern: ^[0-9]+$
          type: string
        created_at:
          type: string
          format: date-time
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PageResponseBookingDTO:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BookingDTO'
        total_elements:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
        current_page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
...
    FieldErrorDTO:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
    ProblemDTO:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDTO'
        trace_id:
          type: string
    RestResponseProblemDTO:
      required:
      - statusCode
      - message
      - data
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
          example: 500
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProblemDTO'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...